# api/openapi/control/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Control API
  version: 1.0.0
  description: V1 HTTP contract for Identity, platform administration, Tenant membership, Agent authoring, Runtime Profile authoring, and Deployment publication.
servers:
  - url: /
...
paths:
...
  /v1/tenants/{tenant_id}/deployments:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    post:
      operationId: createDeployment
      summary: Create a stable Deployment identity
      description: >-
        Creates metadata with revision 1 and no published revision. The
        Idempotency-Key is scoped to this Tenant; an identical replay returns
        the original creation snapshot rather than current mutable metadata.
      tags: [Deployment]
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/DeploymentIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateDeploymentRequest'}
      responses:
        '200':
          description: The same create command was already committed; its original snapshot is returned.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateDeploymentResponse'}
        '201':
          description: A new Deployment identity was created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateDeploymentResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '500': {$ref: '#/components/responses/InternalError'}
    get:
      operationId: listDeployments
      summary: List Deployments in the Tenant
      tags: [Deployment]
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of Deployment metadata.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeploymentPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/tenants/{tenant_id}/deployments/{deployment_id}:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/DeploymentID'
    get:
      operationId: getDeployment
      summary: Get Deployment metadata
      tags: [Deployment]
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: Deployment metadata and latest publication number.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Deployment'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
    patch:
      operationId: updateDeployment
      summary: Compare-and-swap mutable Deployment metadata
      description: >-
        Updates name and/or description against expected_metadata_revision.
        A semantic no-op returns the current object without advancing its CAS.
      tags: [Deployment]
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateDeploymentRequest'}
      responses:
        '200':
          description: Current Deployment metadata after applying the command.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Deployment'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '500': {$ref: '#/components/responses/InternalError'}
...
  /v1/admin/capabilities:
    get:
      operationId: getAdminCapabilities
      summary: Get fixed V1 platform administration capabilities
      tags: [Admin]
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: Active operator capabilities.
          content:
            application/json:
              schema:
                type: object
                required: [capabilities]
                properties:
                  capabilities:
                    type: array
                    items: {type: string}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/admin/operators:
    get:
      operationId: listPlatformOperators
      summary: List active Platform Operators
      tags: [Admin]
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: Active Platform Operators.
          content:
            application/json:
              schema:
                type: object
                required: [operators]
                properties:
                  operators:
                    type: array
                    items: {$ref: '#/components/schemas/Operator'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      operationId: grantPlatformOperator
      summary: Grant platform administration to an existing active user
      tags: [Admin]
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GrantOperatorRequest'}
      responses:
        '201':
          description: Operator grant is active.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Operator'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/admin/operators/{user_id}:
    parameters:
      - $ref: '#/components/parameters/UserID'
    delete:
      operationId: revokePlatformOperator
      summary: Revoke a Platform Operator while preserving at least one operator
      tags: [Admin]
      security: [{cookieAuth: []}]
      responses:
        '204':
          description: Operator grant revoked.
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/admin/users:
    get:
      operationId: listPlatformUsers
      summary: List global platform users
      tags: [Admin]
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of global users.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      operationId: createPlatformUser
      summary: Create an active user with a temporary password
      tags: [Admin]
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateUserRequest'}
      responses:
        '201':
          description: Global user created; password change is required on first login.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/admin/tenants:
    get:
      operationId: listTenantsForOperations
      summary: List Tenants for platform operations
      tags: [Admin]
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of Tenants.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      operationId: provisionTenant
      summary: Create a Tenant and an initial owner membership
      tags: [Admin]
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProvisionTenantRequest'}
      responses:
        '201':
          description: Tenant and initial owner membership created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Tenant'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: control_session
  parameters:
    TenantID:
      name: tenant_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    DeploymentID:
      name: deployment_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    DeploymentIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Stable key for one logical Deployment create or publish command. Reuse
        with an identical payload returns the original fixed result; reuse with
        another payload returns 409. The plaintext key is not persisted.
      schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[!-~]+$'}
...
    UserID:
      name: user_id
      in: path
      required: true
      schema: {type: string, minLength: 1}
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    BadRequest:
      description: Request input is invalid.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Unauthorized:
      description: A valid session is required.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Forbidden:
      description: The authenticated user is not authorized or must change a temporary password.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Conflict:
      description: The requested state change conflicts with current state.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    PayloadTooLarge:
      description: The request body exceeds the fixed endpoint limit.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    InternalError:
      description: The request could not be completed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    CreateDeploymentRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 128}
        description: {type: string, maxLength: 4096}
    UpdateDeploymentRequest:
      type: object
      additionalProperties: false
      minProperties: 2
      required: [expected_metadata_revision]
      properties:
        expected_metadata_revision: {type: integer, minimum: 1}
        name: {type: string, minLength: 1, maxLength: 128}
        description: {type: string, maxLength: 4096}
    Deployment:
      type: object
      additionalProperties: false
      required: [id, tenant_id, name, description, metadata_revision, latest_revision_number, created_by, created_at, updated_at]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        tenant_id: {type: string, minLength: 1, maxLength: 128}
        name: {type: string, minLength: 1, maxLength: 128}
        description: {type: string, maxLength: 4096}
        metadata_revision: {type: integer, minimum: 1}
        latest_revision_number: {type: integer, minimum: 1, nullable: true}
        created_by: {type: string, minLength: 1, maxLength: 128}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateDeploymentResponse:
      type: object
      additionalProperties: false
      required: [deployment]
      properties:
        deployment: {$ref: '#/components/schemas/Deployment'}
    DeploymentPage:
      type: object
      additionalProperties: false
      required: [deployments, offset, limit, total]
      properties:
        deployments:
          type: array
          items: {$ref: '#/components/schemas/Deployment'}
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
...
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, temporary_password]
      properties:
        username: {type: string, minLength: 3, maxLength: 32}
        display_name: {type: string, maxLength: 256}
        temporary_password: {type: string, minLength: 12, maxLength: 128, format: password}
    GrantOperatorRequest:
      type: object
      additionalProperties: false
      required: [user_id]
      properties:
        user_id: {type: string, minLength: 1}
    ProvisionTenantRequest:
      type: object
      additionalProperties: false
      required: [slug, name, owner_user_id]
      properties:
        slug: {type: string, pattern: '^[a-z0-9][a-z0-9-]{2,31}$'}
        name: {type: string, minLength: 1, maxLength: 100}
        owner_user_id: {type: string, minLength: 1}
...
    User:
      type: object
      additionalProperties: false
      required: [id, username, display_name, status, created_at, updated_at]
      properties:
        id: {type: string}
        username: {type: string}
        display_name: {type: string}
        status: {type: string, enum: [ACTIVE, DISABLED]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Operator:
      type: object
      additionalProperties: false
      required: [user_id, granted_at]
      properties:
        user_id: {type: string}
        username: {type: string}
        display_name: {type: string}
        granted_by: {type: string}
        granted_at: {type: string, format: date-time}
    Tenant:
      type: object
      additionalProperties: false
      required: [id, slug, name, status, created_at, updated_at]
      properties:
        id: {type: string}
        slug: {type: string}
        name: {type: string}
        status: {type: string, enum: [ACTIVE]}
        owner_user_id: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    UserPage:
      type: object
      additionalProperties: false
      required: [users, offset, limit, total]
      properties:
        users:
          type: array
          items: {$ref: '#/components/schemas/User'}
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
    TenantPage:
      type: object
      additionalProperties: false
      required: [tenants, offset, limit, total]
      properties:
        tenants:
          type: array
          items: {$ref: '#/components/schemas/Tenant'}
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}